+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - update tooltip to not use parent CSS config
+- update MenuItem to only show focus on keyboard navigation
+- update test to respond to the change
 
 ## [0.7.10]